// apps/ledger/src/components/data-table-filter/core/types.ts
export type ElementType<T> = T extends (infer U)[] ? U : T
// ...
export type ColumnDataType =
	/* The column value is a string that should be searchable. */
	| "text"
	| "number"
	| "date"
	/* The column value can be a single value from a list of options. */
	| "option"
	/* The column value can be zero or more values from a list of options. */
	| "multiOption"

/*
 * Represents the data type (kind) of option and multi-option columns.
 */
export type OptionBasedColumnDataType = Extract<ColumnDataType, "option" | "multiOption">

/*
 * Maps a ColumnDataType to it's primitive type (i.e. string, number, etc.).
 */
export type ColumnDataNativeMap = {
	text: string
	number: number
	date: Date
	option: string
	multiOption: string[]
}

/*
 * Represents the value of a column filter.
 * Contigent on the filtered column's data type.
 */
export type FilterValues<T extends ColumnDataType> = Array<ElementType<ColumnDataNativeMap[T]>>
// ...
export type TextFilterOperator = "contains" | "does not contain"

/* Operators for number data */
export type NumberFilterOperator =
	| "is"
	| "is not"
	| "is less than"
	| "is greater than or equal to"
	| "is greater than"
	| "is less than or equal to"
	| "is between"
	| "is not between"

/* Operators for date data */
export type DateFilterOperator =
	| "is"
	| "is not"
	| "is before"
	| "is on or after"
	| "is after"
	| "is on or before"
	| "is between"
	| "is not between"

/* Operators for option data */
export type OptionFilterOperator = "is" | "is not" | "is any of" | "is none of"

/* Operators for multi-option data */
export type MultiOptionFilterOperator =
	| "include"
	| "exclude"
	| "include any of"
	| "include all of"
	| "exclude if any of"
	| "exclude if all"

/* Maps filter operators to their respective data types */
export type FilterOperators = {
	text: TextFilterOperator
	number: NumberFilterOperator
	date: DateFilterOperator
	option: OptionFilterOperator
	multiOption: MultiOptionFilterOperator
}

/*
 *
 * FilterValue is a type that represents a filter value for a specific column.
 *
 * It consists of:
 * - Operator: The operator to be used for the filter.
 * - Values: An array of values to be used for the filter.
 *
 */
export type FilterModel<TType extends ColumnDataType = any> = {
	columnId: string
	type: TType
	operator: FilterOperators[TType]
	values: FilterValues<TType>
}